add Loader for custom and secure file system access (#17)
Currently, LoadFromFile and LoadFromReader will attempt to open and read
externally linked tilesets using os.Open. This can be an issue if maps
are from an untrusted source.

To solve this, an http.FileSystem implementation can be set in the Loader
that is used when loading maps and externally linked tilesets.