Fix Tilesets not being loaded if only referenced by tile objects.
TMX maps contain object groups that can be of several types, such as,
rectangles, points, etc. One of these types is a tile object, which
are objects that have a tile GID field. If this field is set then the
tile refers to a tile within a tileset.

If the tileset is used by a tile object but otherwise not used in any
of the layers, then the library will not load this tileset.

This patch extends the decoding logic to include decoding object
groups also. To decode an object group, all objects contained in the
object group are traversed and in case the object has a GID set, then
the corresponding tileset is initialized by calling TileGIDToTile,
which is also the stragegy used by layer decoding.