Add "class" attribute to several elements.
In the TMX version 1.9
(https://doc.mapeditor.org/en/stable/reference/tmx-changelog/#tiled-1-9),
the "type" attribute for <tile> and <object> was renamed to "class".

Also, the "class" field was added to several elements, namely, <map>,
<tileset>, <layer>, <imagelayer>, <objectgroup>, <group>, <wangset>
and <wangcolor>.

This patch adds the "class" attribute to all of the elements that are
specified in version 1.9 and that are implemented by the go-tiled
library.