Add getter / setter for Tileset.baseDir.
It's possible to load TSX files directly using the xml.NewDecoder() but it's not
possible to set the Tileset.baseDir field since this field is not exported.

This means that for TSX files loaded via the xml.NewDecoder(), it's not possible
to use the Tileset.GetFileFullPath() method to determine the location of
resources relative to the path of the tileset, e.g., the tileset image.

So far I have been employing a workaround which is to always pass the Tileset
and also the baseDir together as separate arguments, which is cumbersome.

It would be ideal if it were possible to set the Tileset.baseDir field
directly. This means that if a TMX map is loaded by the go-tiled library, then
the go-tiled library sets the Tileset.baseDir. However, if the TSX tileset is
loaded manually, then the code loading it manually can also set the field.

An alternative would be to extend the go-tiled library with functions to load
the Tileset from a file, from a reader, etc, but this would be additional
effort.

This patch adds getter / setter for the Tileset.baseDir field as suggested in
https://github.com/lafriks/go-tiled/pull/67#issuecomment-1214221651.